Added inner auth wall in the departments petitions

+++ backend/src/Components/PetitionResolvers/DepartmentsPetitions.ts
@@ -1,43 +1,64 @@
 import { Router, Request, Response } from "express";
 import Database from "../Database";
+import { auth } from "../Auth/InnerAuthValidator";
 
 let db: Database = new Database();
 const DepartmentsRouter = Router();
 
 
 DepartmentsRouter.post("/editDepartment", async (req: Request, res: Response) => {
-  try {
-    const response = await db.EditDepartment(req.body.name, req.body.department_id);
-    res.status(200).send({response: "OK"});
+  const authed = auth(req);
+  if(authed) {
+    try {
+      const response = await db.EditDepartment(req.body.name, req.body.department_id);
+      res.status(200).send({response: "OK"});
+    }
+    catch(ex) {
+      console.log(ex);
+      res.status(500).send({response: "Server error"});
+    }
   }
-  catch(ex) {
-    console.log(ex);
-    res.status(500).send({response: "Server error"});
+  else {
+    res.status(401).json({ result: "PROHIBITED"});
   }
 });
 
 DepartmentsRouter.post("/hideShowDepartment", async (req: Request, res: Response) => {
-  await db.ChangeVisibility(req.body.department_id);
-  res.status(200).send({response: "OK"});
+  const authed = auth(req);
+  if(authed) {
+    await db.ChangeVisibility(req.body.department_id);
+    res.status(200).send({response: "OK"});
+  }
+  else {
+    res.status(401).json({ result: "PROHIBITED"});
+  }
 });
 
 DepartmentsRouter.post("/newDepartment", async (req: Request, res: Response) => {
-  console.log("Received a request to create a new department named: " + req.body.name);
-  
-  const response = await db.AddDepartment(req.body.name);
-
-  res.status(200).send({response: "OK"});
+  const authed = auth(req);
+  if(authed) {
+    const response = await db.AddDepartment(req.body.name);
+    res.status(200).send({response: "OK"});
+  }
+  else {
+    res.status(401).json({ result: "PROHIBITED"});
+  }
 });
 
 DepartmentsRouter.post("/deleteDepartment", async (req: Request, res: Response) => {
-  console.log(req.body.department_id);
-  try {
-    const response = await db.DeleteDepartment(req.body.department_id);
-    res.status(200).send({response: "OK"});
+  const authed = auth(req);
+  if(authed) {
+    try {
+      const response = await db.DeleteDepartment(req.body.department_id);
+      res.status(200).send({response: "OK"});
+    }
+    catch(ex) {
+      console.log(ex);
+      res.status(500).send({response: "Server error"});
+    }  
   }
-  catch(ex) {
-    console.log(ex);
-    res.status(500).send({response: "Server error"});
+  else {
+    res.status(401).json({ result: "PROHIBITED"});
   }
 });
